+++ tests/testthat/test-textplot.R
@@ -2,6 +2,7 @@ context("test textplot_*")
 
 toks_test <- readRDS("../data/tokens_test.RDS")
 test_toks <- tokens_remove(toks_test, stopwords())
+dict <- dictionary(list("keywords" = c("positive", "bad", "xxxx")))
 
 test_that("textplot_* works", {
     dfmt <- dfm(test_toks)
@@ -9,8 +10,9 @@ test_that("textplot_* works", {
     lss <- textmodel_lss(dfmt, seed, k = 10)
     expect_equal(class(textplot_simil(lss, group = TRUE)), c("gg", "ggplot"))
     expect_equal(class(textplot_simil(lss, group = FALSE)), c("gg", "ggplot"))
-    expect_equal(class(textplot_factor(lss)), c("gg", "ggplot"))
-    expect_equal(class(textplot_terms(lss, highlighted = c("positive", "bad", "xxxx"))),
+    expect_equal(class(textplot_terms(lss, highlighted = dict$keywords)),
+                 c("gg", "ggplot"))
+    expect_equal(class(textplot_terms(lss, highlighted = dict)),
                  c("gg", "ggplot"))
     expect_equal(class(textplot_terms(lss)), c("gg", "ggplot"))
 })
@@ -20,7 +22,8 @@ test_that("textplot_* raise error when attributes are missing", {
     coef <- rnorm(100)
     names(coef) <- topfeatures(dfmt, 100)
     lss <- as.textmodel_lss(coef)
-    expect_error(textplot_simil(lss, group = TRUE), "Invalid textmodel_lss object")
-    expect_error(textplot_simil(lss, group = FALSE), "Invalid textmodel_lss object")
-    expect_error(textplot_factor(lss), "Invalid textmodel_lss object")
+    expect_error(textplot_simil(lss, group = TRUE),
+                 "textplot_simil() does not work with dummy models", fixed = TRUE)
+    expect_error(textplot_simil(lss, group = FALSE),
+                 "textplot_simil() does not work with dummy models", fixed = TRUE)
 })