+++ tests/testthat/test-utils.R
@@ -47,17 +47,16 @@ test_that("as.seedwords works", {
 
 
 test_that("cohesion works", {
-    lis <- cohesion(lss_test)
-    expect_identical(names(lis), c("overall", "component"))
-    expect_identical(names(lis$component), c("k", "raw", "smoothed"))
-    expect_identical(nrow(lis$component), lss_test$k)
+    coh <- cohesion(lss_test)
+    expect_identical(names(coh), c("k", "raw", "smoothed"))
+    expect_identical(nrow(coh), lss_test$k)
 })
 
 test_that("strength works", {
     lis <- strength(lss_test)
     expect_identical(names(lis), c("overall", "element"))
     expect_identical(names(lis$element), c("seed", "selected", "all"))
-    expect_identical(nrow(lis$element), length(unlist(lss_test$seeds)))
+    expect_identical(nrow(lis$element), length(unlist(lss_test$seeds_weighted)))
 
     lss1 <- textmodel_lss(dfmt_test, seed_test, terms = feat_test, k = 300, slice = 100)
     lss2 <- textmodel_lss(dfmt_test, seed_test, terms = feat_test, k = 300)